+++ src/main/java/com/ciphertext/opencarebackend/dto/response/BloodDonationResponse.java
@@ -3,6 +3,7 @@ package com.ciphertext.opencarebackend.dto.response;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -15,10 +16,10 @@ public class BloodDonationResponse {
     private Long quantityMl;
     private String donationType;
     private String healthStatus;
-    private Double hemoglobinLevel;
+    private BigDecimal hemoglobinLevel;
     private String bloodPressure;
     private Long pulseRate;
-    private Double temperature;
+    private BigDecimal temperature;
     private String notes;
     private String status;
 }
